requirements.txt
streamlit
setuptools>=80.9.0
numpy>=2.3.4
torch>=2.9.0
matplotlib>=3.10.7
pillow>=12.0.0
torchvision>=0.24.0
kagglehub>=0.3.13